src/screens/Register.js: moved back to screens dir
